pyproject.toml
[project]
name = "simple-cv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.8",
    "pandas>=3.0.1",
    "pillow>=12.1.1",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.8.0",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
]

